divide the sva script into two (one for monitoring one for waveforms) & config json files & other comments
