+++ rnog_analysis_tools/data_monitoring/science_verification_analysis/config_files_sva/config_block_offsets.json
@@ -1,0 +1,6 @@
+{
+    "block_offset_limits": {
+        "median": 0.002,
+        "iqr": 0.005
+    }
+}
