--- rnog_analysis_tools/data_monitoring/science_verification_analysis/config_files_sva/config_block_offsets.py
@@ -1,4 +1,0 @@
-block_offset_limits = {
-    "median": 0.002,
-    "iqr": 0.005
-}
