--- rnog_analysis_tools/data_monitoring/science_verification_analysis/config_files_sva/config_glitching.py
@@ -1,7 +1,0 @@
-config_glitching_values = {
-    "alpha": 0.01,
-    "pvalue": 0.1,
-    "ci_level": 0.99,
-    "strong_ci_threshold": 0.3,
-    "moderate_ci_threshold": 0.2,
-    }
