+++ rnog_analysis_tools/data_monitoring/science_verification_analysis/config_files_sva/config_station.py
@@ -33,5 +33,3 @@ def get_station_config(station_id):
         cfg.update(adjustments)
     return cfg
 
-sampling_rate = {"after_2024": 2.4*units.GHz,
-                 "before_2024": 3.2*units.GHz}
