+++ rnog_analysis_tools/data_monitoring/science_verification_analysis/helper_functions/config_helper.py
@@ -1,0 +1,6 @@
+def get_station_config(station_id, default_station_config, station_specific_adjustments):
+    cfg = default_station_config.copy()
+    if str(station_id) in station_specific_adjustments:
+        adjustments = station_specific_adjustments[str(station_id)]
+        cfg.update(adjustments)
+    return cfg
