+++ rnog_analysis_tools/data_monitoring/science_verification_analysis/helper_functions/read_rnog_runtable.py
@@ -1,0 +1,8 @@
+import rnog_data.runtable as rt
+
+#### Runtable query to get run numbers for a given station and time range
+def read_rnog_runtable(station_id: int, start_time: str, stop_time: str):
+    '''Get run numbers from the runtable tool for a given station and time range.'''
+    RunTable = rt.RunTable()
+    testrt = RunTable.get_table( start_time=start_time, stop_time=stop_time, stations=[station_id], run_types = ['physics'])
+    return testrt
